+++ src/ncurseswwinerror/ncurseswwinerror.rs
@@ -37,7 +37,6 @@ pub NCurseswWinError
     StartColorAlreadyCalled = "ncurseswwin::start_color() already called",
     StartColorNotCalled = "ncurseswwin::start_color() not called",
     MaximumRipoffLines { number: usize } = @{ format!("attempt to initialise ripoff {}, maximum ripoff's allowed {}", number, MAX_RIPOFF_LINES) },
-    RipoffNotInitialized { number: usize } = "ripoff line {number} has not been initialised",
     InternalError = "an internal error has occured",
     Panic { message: String } = "{message}",
     OutOfMemory { func: String } = "{func}() out of memory!!!",